#17: Move common loader logic to base loader
